--- packages/openactive-broker-microservice/src/clustertest1/workerThreadsWorker.js
@@ -1,4 +1,0 @@
-const { workerData, parentPort } = require('worker_threads');
-
-const result = Math.floor(Math.random() * workerData);
-parentPort.postMessage(result);
