+++ packages/openactive-broker-microservice/src/state.js
@@ -10,7 +10,7 @@ const { OnePhaseListeners } = require('./onePhaseListeners');
 
 /**
  * @typedef {import('./models/core').FeedContext} FeedContext
- * @typedef {import('./validator/async-validator')} AsyncValidatorWorker
+ * @typedef {import('./validator/old/async-validator')} AsyncValidatorWorker
  */
 /**
  * @typedef {object} PendingResponse
