+++ packages/openactive-broker-microservice/src/validator/types.d.ts
@@ -1,0 +1,17 @@
+export type ValidatorWorkerResponse = {
+  errors: {
+    opportunityId: string;
+    error: unknown[];
+  }[];
+  numItemsPerFeed: {
+    [feedContextIdentifier: string]: number;
+  };
+};
+
+export type ValidatorWorkerRequestParsedItem = {
+  validationMode: string;
+  item: unknown;
+  feedContextIdentifier: string;
+};
+
+export type ValidatorWorkerRequestParsed = ValidatorWorkerRequestParsedItem[];
