+++ packages/openactive-broker-microservice/src/validator/validator-worker.js
@@ -1,46 +1,68 @@
-// workers/auth.js
 const { validate } = require('@openactive/data-model-validator');
-const { expose } = require('threads/worker');
+const { execPipe, filter, toArray, map } = require('iter-tools');
+const { workerData, parentPort } = require('worker_threads');
 const { silentlyAllowInsecureConnections } = require('../util/suppress-unauthorized-warning');
+const { VALIDATOR_TMP_DIR } = require('../broker-config');
 
-// Note this is duplicated between app.js and validator.js, for efficiency
-const VALIDATOR_TMP_DIR = './tmp';
+/**
+ * @typedef {import('./types').ValidatorWorkerRequestParsed} ValidatorWorkerRequestParsed
+ * @typedef {import('./types').ValidatorWorkerResponse} ValidatorWorkerResponse
+ */
 
 silentlyAllowInsecureConnections();
 
-/**
- * Use OpenActive validator to validate the RPDE item
- *
- * @param {any} body the data item
- */
-async function validateItem(body, validationMode) {
-  /* When re-harvesting the feed frequently during development, this can speed up the process. However, note
-  that leaving this on may allow Broker to miss some critical issues which will cause confusing errors later down
-  the line */
-  if (process.env.DEBUG_BROKER_NO_VALIDATE === 'true') {
-    return [];
+async function run() {
+  // JSON parsing is included in the validatorWorker as it's CPU intensive
+  /** @type {ValidatorWorkerRequestParsed} */
+  const requestParsed = JSON.parse(workerData);
+  /** @type {ValidatorWorkerResponse['numItemsPerFeed']} */
+  const numItemsPerFeed = {};
+  /** @type {ValidatorWorkerResponse['errors']} */
+  const errors = [];
+  for (const { feedContextIdentifier, validationMode, item } of requestParsed) {
+    numItemsPerFeed[feedContextIdentifier] = (numItemsPerFeed[feedContextIdentifier] ?? 0) + 1;
+
+    const allOaValidationErrors = await validate(item, {
+      loadRemoteJson: true,
+      remoteJsonCachePath: VALIDATOR_TMP_DIR,
+      remoteJsonCacheTimeToLive: 3600,
+      validationMode,
+    });
+    const newErrors = execPipe(allOaValidationErrors,
+      filter((oaValidationError) => (
+        oaValidationError.severity === 'failure'
+        && !isIgnorableResult(item, oaValidationError)
+      )),
+      map(((oaValidationError) => ({
+        opportunityId: item['@id'],
+        error: oaValidationError,
+      }))),
+      toArray);
+    errors.push(...newErrors);
   }
-  /**
-   * @type {{
-   *   loadRemoteJson: boolean,
-   *   remoteJsonCachePath: string,
-   *   remoteJsonCacheTimeToLive: number,
-   *   validationMode?: string,
-   * }}
-   */
-  const optionsWithRemoteJson = {
-    loadRemoteJson: true,
-    remoteJsonCachePath: VALIDATOR_TMP_DIR,
-    remoteJsonCacheTimeToLive: 3600,
-    validationMode,
+  /** @type {ValidatorWorkerResponse} */
+  const response = {
+    errors,
+    numItemsPerFeed,
   };
-
-  const errors = (await validate(body, optionsWithRemoteJson))
-    .filter((result) => result.severity === 'failure');
-
-  return errors;
+  parentPort.postMessage(response);
 }
 
-expose({
-  validateItem,
-});
+/**
+ * @param {any} item
+ * @param {any} oaValidationResult
+ * @returns {boolean}
+ */
+function isIgnorableResult(item, oaValidationResult) {
+  // Ignore the error that a SessionSeries must have children as they haven't been combined yet.
+  // This is being done because I don't know if there is a validator.validationMode for this, and without ignoring the broker does not run
+  if (item['@type'] === 'SessionSeries' && oaValidationResult.message.startsWith('A `SessionSeries` must have an `eventSchedule` or at least one `subEvent`.')) {
+    return true;
+  }
+  return false;
+}
+
+run();
+
+// This is just required in order to use the ValidatorWorkResponse type in other files
+module.exports = {};
